leave input STARTED during failover recovery and change trace log message to warn level
Don't change running state to STARTING when the connection is disrupted
after all, as this had unwanted side effects and interfered with
recovery after an outage. Instead, only the log warning messages are
left in.

Sideline: Also changed a trace log message to warn level for a consumed
JMS "Stopped." exception.