failover thread synchronization / resource cleanup
Improvements to thread synchronization and resource cleanup, made while
trying to fix the lock-up that occurred if the failover or discovery
transports are used and there are no available brokers. (Not certain if
the fix is 100% for all cases where that bug occurred, but the locking
behavior appears to be greatly reduced, at least) In addition, warnings
are written to the log when the failover or discovery transport
reconnects to a different server.