Change choice of Queue vs. Topic to a pick list
Instead of requiring users to type in the exact strings "Queue" /
"Topic" for the Destination Type, change the Transport Definitions to
present a pick list of allowed values. (Caveat: if an existing input
used a different case, e.g. "queue", the currrent value might be lost
and reset to default the next time you edit the input in GeoEvent
Manager. Check any existing inputs after deploying.)