Update ActiveMQ Client to 5.15.4 and Bundle Plugin to 3.5.0
Updated these dependencies before troubleshooting an issue with this component failing to initialize properly on 10.6. This dependency changes already seem to have it working with 10.6 again. I believe it was the Bundle Plugin that made the difference. (Note, it is built against GeoEvent SDK 10.4.1 and the transport still reports as version 10.3.0 for backward compatibility.)